+++ spec/02-integration/05-proxy/25-upstream_keepalive_spec.lua
@@ -74,6 +74,27 @@ describe("#postgres upstream keepalive", function()
       },
     }
 
+    -- crc32 collision upstream TLS
+    bp.routes:insert {
+      hosts = { "plumless.com" },
+      preserve_host = true,
+      service = bp.services:insert {
+        protocol = helpers.mock_upstream_ssl_protocol,
+        host = helpers.mock_upstream_hostname,
+        port = helpers.mock_upstream_ssl_port,
+      },
+    }
+
+    bp.routes:insert {
+      hosts = { "buckeroo.com" },
+      preserve_host = true,
+      service = bp.services:insert {
+        protocol = helpers.mock_upstream_ssl_protocol,
+        host = helpers.mock_upstream_hostname,
+        port = helpers.mock_upstream_ssl_port,
+      },
+    }
+
     -- upstream mTLS
     bp.routes:insert {
       hosts = { "example.com", },
@@ -124,6 +145,17 @@ describe("#postgres upstream keepalive", function()
           .has
           .line([[enabled connection keepalive \(pool=[A-F0-9.:]+\|\d+\|one.com]])
 
+    assert.errlog()
+          .has.line([[keepalive get pool, name: [A-F0-9.:]+\|\d+\|one.com, cpool: 0+]])
+    assert.errlog()
+          .has.line([[keepalive create pool, name: [A-F0-9.:]+\|\d+\|one.com, size: \d+]])
+    assert.errlog()
+          .has.line([[lua balancer: keepalive no free connection, cpool: [A-F0-9]+]])
+    assert.errlog()
+          .has.line([[lua balancer: keepalive saving connection [A-F0-9]+, cpool: [A-F0-9]+]])
+    assert.errlog()
+          .not_has.line([[keepalive free pool]], true)
+
     local res = assert(proxy_client:send {
       method = "GET",
       path = "/echo_sni",
@@ -136,6 +168,34 @@ describe("#postgres upstream keepalive", function()
     assert.errlog()
           .has
           .line([[enabled connection keepalive \(pool=[A-F0-9.:]+\|\d+\|two.com]])
+
+    assert.errlog()
+          .has.line([[keepalive get pool, name: [A-F0-9.:]+\|\d+\|two.com, cpool: 0+]])
+    assert.errlog()
+          .has.line([[keepalive create pool, name: [A-F0-9.:]+\|\d+\|two.com, size: \d+]])
+    assert.errlog()
+          .has.line([[lua balancer: keepalive no free connection, cpool: [A-F0-9]+]])
+    assert.errlog()
+          .has.line([[lua balancer: keepalive saving connection [A-F0-9]+, cpool: [A-F0-9]+]])
+    assert.errlog()
+          .not_has.line([[keepalive free pool]], true)
+
+    local handle, result
+
+    handle = io.popen([[grep 'cpool: 0000000000' servroot/logs/error.log|wc -l]])
+    result = handle:read("*l")
+    handle:close()
+    assert(tonumber(result) == 2)
+
+    handle = io.popen([[grep 'keepalive create pool, name:' servroot/logs/error.log|wc -l]])
+    result = handle:read("*l")
+    handle:close()
+    assert(tonumber(result) == 2)
+
+    handle = io.popen([[grep 'lua balancer: keepalive saving connection' servroot/logs/error.log|wc -l]])
+    result = handle:read("*l")
+    handle:close()
+    assert(tonumber(result) == 2)
   end)
 
 
@@ -165,8 +225,18 @@ describe("#postgres upstream keepalive", function()
     assert.not_equal(fingerprint_1, fingerprint_2)
 
     assert.errlog()
-              .has
-              .line([[enabled connection keepalive \(pool=[0-9.]+|\d+|[0-9.]+:\d+|[a-f0-9-]+]])
+          .has.line([[enabled connection keepalive \(pool=[0-9.]+|\d+|[0-9.]+:\d+|[a-f0-9-]+]])
+    assert.errlog()
+          .has.line([[keepalive get pool, name: [0-9.]+|\d+|[0-9.]+:\d+|[a-f0-9-]+, cpool: 0+]])
+    assert.errlog()
+          .has.line([[keepalive create pool, name: [0-9.]+|\d+|[0-9.]+:\d+|[a-f0-9-]+, size: \d+]])
+    assert.errlog()
+          .has.line([[lua balancer: keepalive no free connection, cpool: [A-F0-9]+]])
+    assert.errlog()
+          .has.line([[lua balancer: keepalive saving connection [A-F0-9]+, cpool: [A-F0-9]+]])
+
+    assert.errlog()
+          .not_has.line([[keepalive free pool]], true)
   end)
 
 
@@ -188,6 +258,11 @@ describe("#postgres upstream keepalive", function()
           .not_has
           .line("enabled connection keepalive", true)
 
+    assert.errlog()
+          .not_has.line([[keepalive get pool]], true)
+    assert.errlog()
+          .not_has.line([[keepalive create pool]], true)
+
     local res = assert(proxy_client:send {
       method = "GET",
       path = "/echo_sni",
@@ -200,6 +275,159 @@ describe("#postgres upstream keepalive", function()
     assert.errlog()
           .not_has
           .line("enabled connection keepalive", true)
+
+    assert.errlog()
+          .not_has.line([[keepalive get pool]], true)
+    assert.errlog()
+          .not_has.line([[keepalive create pool]], true)
+    assert.errlog()
+          .not_has.line([[keepalive free pool]], true)
   end)
+
+
+  it("reuse upstream keepalive pool", function()
+    start_kong()
+
+    local res = assert(proxy_client:send {
+      method = "GET",
+      path = "/echo_sni",
+      headers = {
+        Host = "one.com",
+      }
+    })
+    local body = assert.res_status(200, res)
+    assert.equal("SNI=one.com", body)
+    assert.errlog()
+          .has
+          .line([[enabled connection keepalive \(pool=[A-F0-9.:]+\|\d+\|one.com]])
+
+    assert.errlog()
+          .has.line([[keepalive get pool, name: [A-F0-9.:]+\|\d+\|one.com, cpool: 0+]])
+    assert.errlog()
+          .has.line([[keepalive create pool, name: [A-F0-9.:]+\|\d+\|one.com, size: \d+]])
+    assert.errlog()
+          .has.line([[keepalive no free connection, cpool: [A-F0-9]+]])
+    assert.errlog()
+          .has.line([[keepalive saving connection [A-F0-9]+, cpool: [A-F0-9]+]])
+    assert.errlog()
+          .not_has.line([[keepalive free pool]], true)
+
+    local handle, upool_ptr
+
+    handle = io.popen([[grep 'lua balancer: keepalive saving connection' servroot/logs/error.log]] .. "|" ..
+                      [[grep -Eo 'cpool: [A-F0-9]+']])
+    upool_ptr = handle:read("*l")
+    handle:close()
+
+    local res = assert(proxy_client:send {
+      method = "GET",
+      path = "/echo_sni",
+      headers = {
+        Host = "one.com",
+      }
+    })
+    local body = assert.res_status(200, res)
+    assert.equal("SNI=one.com", body)
+    assert.errlog()
+          .has
+          .line([[enabled connection keepalive \(pool=[A-F0-9.:]+\|\d+\|one.com]])
+
+    assert.errlog()
+          .has.line([[keepalive get pool, name: [A-F0-9.:]+\|\d+\|one.com, ]] .. upool_ptr)
+    assert.errlog()
+          .has.line([[keepalive reusing connection [A-F0-9]+, requests: \d+, ]] .. upool_ptr)
+    assert.errlog()
+          .has.line([[keepalive saving connection [A-F0-9]+, ]] .. upool_ptr)
+    assert.errlog()
+          .not_has.line([[keepalive free pool]], true)
+  end)
+
+
+  it("free upstream keepalive pool", function()
+    start_kong({ upstream_keepalive_max_requests = 1, })
+
+    local res = assert(proxy_client:send {
+      method = "GET",
+      path = "/echo_sni",
+      headers = {
+        Host = "one.com",
+      }
+    })
+    local body = assert.res_status(200, res)
+    assert.equal("SNI=one.com", body)
+    assert.errlog()
+          .has
+          .line([[enabled connection keepalive \(pool=[A-F0-9.:]+\|\d+\|one.com]])
+
+    assert.errlog()
+          .has.line([[keepalive get pool, name: [A-F0-9.:]+\|\d+\|one.com, cpool: 0+]])
+    assert.errlog()
+          .has.line([[keepalive create pool, name: [A-F0-9.:]+\|\d+\|one.com, size: \d+]])
+    assert.errlog()
+          .has.line([[keepalive no free connection, cpool: [A-F0-9]+]])
+    assert.errlog()
+          .has.line([[keepalive not saving connection [A-F0-9]+, cpool: [A-F0-9]+]])
+    assert.errlog()
+          .has.line([[keepalive free pool, name: [A-F0-9.:]+\|\d+\|one.com, cpool: [A-F0-9]+]])
+
+    assert.errlog()
+          .not_has.line([[keepalive saving connection]], true)
+  end)
+
+
+  it("pools with crc32 collision", function()
+    start_kong()
+
+    local res = assert(proxy_client:send {
+      method = "GET",
+      path = "/echo_sni",
+      headers = {
+        Host = "plumless.com",
+      }
+    })
+    local body = assert.res_status(200, res)
+    assert.equal("SNI=plumless.com", body)
+    assert.errlog()
+          .has
+          .line([[enabled connection keepalive \(pool=[A-F0-9.:]+\|\d+\|plumless.com]])
+
+    local res = assert(proxy_client:send {
+      method = "GET",
+      path = "/echo_sni",
+      headers = {
+        Host = "buckeroo.com",
+      }
+    })
+    local body = assert.res_status(200, res)
+    assert.equal("SNI=buckeroo.com", body)
+    assert.errlog()
+          .has
+          .line([[enabled connection keepalive \(pool=[A-F0-9.:]+\|\d+\|buckeroo.com]])
+
+    local handle
+
+    handle = io.popen([[grep 'enabled connection keepalive ' servroot/logs/error.log]] .. "|" ..
+                      [[grep -Eo 'pool=[A-F0-9.:]+\|\d+\|plumless.com']])
+    local name1 = handle:read("*l")
+    handle:close()
+
+    handle = io.popen([[grep 'enabled connection keepalive ' servroot/logs/error.log]] .. "|" ..
+                      [[grep -Eo 'pool=[A-F0-9.:]+\|\d+\|buckeroo.com']])
+    local name2 = handle:read("*l")
+    handle:close()
+
+    local crc1 = ngx.crc32_long(name1)
+    local crc2 = ngx.crc32_long(name2)
+    assert.equal(crc1, crc2)
+
+    handle = io.popen([[grep 'lua balancer: keepalive saving connection' servroot/logs/error.log]] .. "|" ..
+                      [[grep -Eo 'cpool: [A-F0-9]+']])
+    local upool_ptr1 = handle:read("*l")
+    local upool_ptr2 = handle:read("*l")
+    handle:close()
+
+    assert.not_equal(upool_ptr1, upool_ptr2)
+  end)
+
 
 end)